Completed the automatic stand selection method
After a large amount of trail and error with various methods, the
stratifiedSample function is able to be used to extract 500 points, 50
for each ecoregion utilizing the fast change mask and landcover mask.
This method uses the batch feature collection export to google drive so
the output points are external to the script in a shapefile

This commit still contains many extrainous code blocks that need to be
cleaned out.

Additionally this commit doesn't have a method for reintroducing the
points to be used by methods further down in the script (i.e) value
extraction