Complete Revision of DBEST methods
Changed to only used the change detection algorithm, changed to
change.magnitude setting to have DBEST return multiple segments instead
of just two.

Wrote new function choose the best segments for creation of recovery
metrics as well as calculate recovery metrics and index information for
the segments to respective dataframes. There are a lot of rules to
remove erronous segments and time series as well as combine segments
that should be a single segment (during recovery period)

Changed how plotting happens to only visualize the chosen segments out
of the total. Also plotting now pulls from the segmentation dataframe
and the metrics dataframe instead of trying to calculate everything
again.

Added more recovery metrics and added some calculations at the end to
show remaining bad time series from two sources : ones that could not be
segmented due to rules, and ones which have an unrealistic amount of
time between the distirubance segment and the recovery segment