A complete overhaul of the DBEST segmentation
This revision encompasses a lot of changes to how the time series are
segmented, how recovery metrics are calculated, and how both the
segments and recovery metrics are calculated

The DBEST parameter change.magnitude has been changed in favor of
breakpoint.no to force DBEST to identify change segments even if
the algorithm doesn't find them to meet some criteria

This change allows for a secondary layer of segmentation to happen
based on logic operators that ensure that when recovery metrics are
being calculated they are based on segments that make practical sense,
aka a disturbance should happen before a recovery, there shouldn't be a
gap of more than 3 years between disturbance and recovery etc. Lastly,
index and length information is exported to a seperate auxiliary
dataframe to be used when plotting.

Recovery metrics are now calculated in a function with the additional
layer of segmentation. Additional recovery metrics have been added.
Recovery metrics are now directly accessed by the plotting function
rather than a seperate and duplicate calculation being performed during
plotting.

Lastly, after the main segmentation and plotting occurs there are some
operations to assess how many time series are 'useable' from a realistic
perspective, and some plotting operations are available but these are
not finalized and should have revisions at some point.