Misc updates, backing up before new changes
Not really sure what changes exist, I'm about to make big changes to the
python script so this will serve as a major checkpoint to return to if
anything gets messed up in the process.

Additionally the R script has been updated to include automatic
statistics plots to be created using ggplot2, however these plots are
not finished and will need to be updated with better titles and legends.
The R script also contains a new function and commented out section just
before the DBEST segmentation which allows for an evaluation of
sensativity of the segmentation parameters.