Environmental variable extraction
Python : multiple assets are imported for the extraction of a series of
environmental variables associated with each stand. Data accessed are a
10m DEM which is used to create elevation, aspect, and slope, a soil
texture dataset, a climate
dataset from NASA which projects out to 2099 under different climate
scenarios and provides precipitation, minimum temperature, and maximum
temperature on a monthly basis. These values are extracted as a time
series with the same length as the vegetation index time series and are available for four
different climate scenarios resulting in 12 dataframes being exported. All values are extracted in seperate
code blocks allowing for the individual extraction of desired attributes
without having to wait for others already extracted. Also correctly
formats and exports each of these data types to csv files

R : all environmental varibles are imported, however they have not been
integrated with the segmentation outputs yet as the climate data will
need more processing before it's able to be integrated for analysis.